requirements.txt
alabaster==0.7.12
allennlp==0.9.0
attrs==19.3.0
Babel==2.7.0
blis==0.2.4
boto==2.49.0
boto3==1.10.45
botocore==1.13.45
certifi==2019.11.28
chardet==3.0.4
Click==7.0
conllu==1.3.1
cycler==0.10.0
cymem==2.0.3
dill==0.3.1.1
docutils==0.15.2
editdistance==0.5.3
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gensim==3.8.1
gevent==1.4.0
greenlet==0.4.15
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonnet==0.14.0
jsonpickle==1.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.0
numpydoc==0.9.2
overrides==2.6
packaging==19.2
pandas==0.25.3
parsimonious==0.8.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
protobuf==3.11.2
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.6
pyrouge==0.1.3
pytest==5.3.2
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
PyYAML==5.3
regex==2019.12.20
requests==2.22.0
responses==0.10.9
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
sentencepiece==0.1.85
six==1.13.0
smart-open==1.9.0
snowballstemmer==2.0.0
spacy==2.1.9
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
tensorboardX==1.9
thinc==7.0.8
torch==1.3.1
tqdm==4.41.0
Unidecode==1.1.1
urllib3==1.25.7
wasabi==0.4.2
wcwidth==0.1.7
Werkzeug==0.16.0
word2number==1.1
zipp==0.6.0
